Adjusted scaling of various terms
Main one: Scellier code was not multiplying cost function by .5 in his code, as he was in his paper.

Additional: Scellier was not multiplying quadratic terms by .5 in computation of energy, which might matter if autograd is used but does not matter in current code configuration since energy is not used to compute gradient.