Experimental weight initialization
Inspired by Glorot-Bengio initialization, but small-world connections into a layer are counted as part of 'in' neurons.
Sweeping initialization for intralayer connection weights, as there does not appear to be a clear approachfor them inspired by the Glorot-Bengio initialization. Inversely proportional to number of neurons in layer, to keep amount of information to given neuron in given layer from other neurons in same layer, independent of number of neurons in the layer.